Document and extend BoringWindows to support multiple sources of boring.
Ignore-this: 7375c8912ede6a6a44db4a4b91ffbc33

The Replace and Merge messages are added to support layouts sending a list of
windows that should be skipped over. The sources are tagged by a string key, so
it is possible though unlikely for different sources of boring windows to
interfere with eachother.

darcs-hash:20090406041301-1499c-a7496446b134bfbf4ab867bf110a7386559bb3cc