Changing behaviour of ppUrgent with X.H.DynamicLog
Ignore-this: 3882f36d5c49e53628485c1570bf136a

Currently, the ppUrgent method is an addition to the ppHidden method.
This doesn't make any sense since it is in fact possible to get urgent
windows on the current and visible screens. So I've raised the ppUrgent
printer to be above ppCurrent/ppVisible and dropped its dependency on
ppHidden.

In addition to that this makes it a lot more easier to define a more
custom ppUrgent printer, since you don't have to "undo" the ppHidden
printer anymore. This also basicly removes the need for dzenStrip,
although I just changed the description.

-- McManiaC / Nils

darcs-hash:20090910010411-e34a6-d8ba596b105db26324d799ee72658ec97648df65