Additional messages for SplitGrid layout
Ignore-this: eb945168d1c420e5a9ed87da12a7acf8

This patch introduces two new message SetMasterRows and SetMasterCols for the
X.GridVariants.SplitGrid layout, which set the number of rows/columns in the
master grid to the given value.  This is useful when setting the number of rows
and/or columns non-incrementally using an interface such as GridSelect.

darcs-hash:20091215192142-18a2b-ab6e52541551341405cba967a775252e83d4ce7f