X.A.Navigation2D
Ignore-this: 3860cc71bfc08d99bd8279c2e0945186

This is a new module to support directional navigation across multiple screens.
As such it is related to X.A.WindowNavigation and X.L.WindowNavigation, but it
is more general.  For a detailed discussion of the differences, see
http://www.cs.dal.ca/~nzeh/xmonad/Navigation2D.pdf.

darcs-hash:20111208205842-18a2b-588f34e94f0402a7153d5e9cae60edd322b87c1b