add 'withNthWorkspace' to DynamicWorkspaceOrder.
Ignore-this: f5f87ffe9ddf1a12fab775e6fb8e856f
Note this is very similar to the function of the same name exported by
DynamicWorkspaces.  Ultimately it would probably be cleaner to
generalize the one in DynamicWorkspaces to accept an arbitrary
workspace sort as a parameter; this is left as an exercise for future
hackers.

darcs-hash:20120407184640-b9182-99b3857ef415af2bc9a6bbbc22388145e0145126